python_script_for_cloud/requirements.txt
gspread==5.12.0
oauth2client==4.1.3
pytz==2024.1
Pillow==10.2.0
numpy==1.26.4
matplotlib==3.8.3
requests==2.31.0
PyGithub==2.3.0
